pyproject.toml

[project]
name = "my-assistant"
version = "0.1.0"
description = "Pipeline assistant for email classification, draft generation, and Slack notification."
requires-python = ">=3.9"
dependencies = [
    "langgraph==0.4.3",
    "langchain",
    "langchain_core",
    "langchain_openai",
    "pydantic==2.11.4",
    "slack-sdk>=3.20.2",
    "google-api-python-client>=2.50.0",
    "google-auth-httplib2>=0.1.0",
    "google-auth-oauthlib>=1.0.0",
    "cassandra-driver>=3.25.0",
    "astrapy",
    "openai",
    "requests",
    "python-dotenv",
    "numpy",
    "fastapi>=0.115.12",
    "uvicorn>=0.34.2",
    "python-multipart>=0.0.20",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
